pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["kalilovm <marlenkalilov86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.27"
fastapi = {extras = ["all"], version = "^0.109.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
